+++ test/resources/test_env/setup_oracle.sql
@@ -18,7 +18,7 @@
 -- See the License for the specific language governing permissions and
 -- limitations under the License.
 
-CREATE PLUGGABLE DATABASE test ADMIN USER test IDENTIFIED BY L3tsT3stTh1s++ FILE_NAME_CONVERT=('pdbseed','test');
+CREATE PLUGGABLE DATABASE test ADMIN USER test IDENTIFIED BY "L3tsT3stTh1s++" FILE_NAME_CONVERT=('pdbseed','test');
 ALTER PLUGGABLE DATABASE test OPEN;
 ALTER PLUGGABLE DATABASE test SAVE STATE;
 ALTER SESSION SET CONTAINER=test;